Remove shadows and some dead code
