+++ Radis/Assets/Editor/imgui.ini
@@ -60,6 +60,7 @@ Size=1280,688
 Collapsed=0
 
 [Window][Entities##window]
+Pos=0,32
 Size=262,688
 Collapsed=0
 DockId=0x00000002,0
