--- Radis/Assets/Shaders/msmBlurV.comp
@@ -1,45 +1,0 @@
-#version 460
-layout(local_size_x=16, local_size_y=16) in;
-
-layout(set=0,binding=0) uniform sampler2D src;
-layout(set=0,binding=1, rgba32f) uniform writeonly image2D dst;
-
-layout(push_constant) uniform PC {
-    int radius;
-    int width;
-    int height;
-} pc;
-
-const int MAX_R = 64;
-shared vec4 tile[16 + 2*MAX_R][16];
-
-void main()
-{
-    ivec2 gid = ivec2(gl_GlobalInvocationID.xy);
-    ivec2 lid = ivec2(gl_LocalInvocationID.xy);
-
-    int R     = clamp(pc.radius, 0, MAX_R);
-    int tileH = 16 + 2*R;
-
-    // All threads participate in loading — no early exit before barrier
-    for (int ty = lid.y; ty < tileH; ty += 16)
-    {
-        int sy = int(gl_WorkGroupID.y)*16 + (ty - R);
-        sy = clamp(sy, 0, pc.height-1);
-        tile[ty][lid.x] = texelFetch(src, ivec2(gid.x, sy), 0);
-    }
-
-    barrier();
-
-    // Out-of-bounds guard is AFTER the barrier
-    if (gid.x >= pc.width || gid.y >= pc.height) return;
-
-    vec4 sum   = vec4(0.0);
-    int center = lid.y + R;
-
-    // Box filter: uniform weight 1 for every tap
-    for (int i = -R; i <= R; ++i)
-        sum += tile[center + i][lid.x];
-
-    imageStore(dst, gid, sum / float(2*R + 1));
-}
