+++ Radis/Radis.vcxproj.user
@@ -16,6 +16,6 @@
     <LocalDebuggerCommandArgumentsHistory>$DevBuild$|</LocalDebuggerCommandArgumentsHistory>
   </PropertyGroup>
   <PropertyGroup>
-    <ShowAllFiles>false</ShowAllFiles>
+    <ShowAllFiles>true</ShowAllFiles>
   </PropertyGroup>
 </Project>
