+++ Radis/src/Radis/Graphics/Vulkan/Uniform/ShaderTypes.h
@@ -79,28 +79,6 @@ namespace Radis
         float texU, texV;
     };
 
-    struct ShadowCameraUniform {
-        glm::mat4 lightViewProj;
-        glm::mat4 lightView;
-        float z0;
-        float z1;
-        float pad0 = 7777.0f;
-        float pad1 = 7777.0f;
-    };
-
-    struct ShadowParamsUniform {
-        glm::mat4 lightViewProj;
-        glm::mat4 lightView;
-        glm::vec4 zParams;     // (z0, z1, 1/(z1-z0), alpha)
-        glm::vec4 mapParams;   // (invW, invH, blurRadius, pad)
-    };
-
-    struct MSMBlurPC {
-        int radius;
-        int width;
-        int height;
-    };
-
     static uint32_t PackColor(const glm::vec3& c)
     {
         return (uint32_t(glm::clamp(c.r, 0.f, 1.f) * 255.f))
